Add markdown README for GitHub
This contains instructions how to build the distribution from source and how
to run the test suite.  I'm aware that a README exists in the distribution
tarball, however this adds a README to GitHub and thus gives users some
information about the project when they see it on GitHub and how they can go
about installing, building, and testing the distribution.